// src/lib/setting/types.ts
// src/lib/setting/types.ts
export interface SettingArchetype {
  name: string;
  theoreticalBasis: string;
  coreDescription: string;
  spatialMechanism: string;
  sensoryHierarchy: string;
  characterReveal: string;
  writingDirectives: string[];
  failureModes: string[];
  systemDirectives: string[];
  writingNotes: string;
}